Update Rust toolchain to support Rust 1.86 (`nightly-2025-02-16`)

+++ libs/core/src/primitive.rs
@@ -12,7 +12,7 @@
 //!     const SOME_PROPERTY: bool = true;
 //! }
 //!
-//! # trait QueryId { const SOME_PROPERTY: core::primitive::bool; }
+//! # trait QueryId { const SOME_PROPERTY: ::core::primitive::bool; }
 //! ```
 //!
 //! Note that the `SOME_PROPERTY` associated constant would not compile, as its
@@ -25,11 +25,17 @@
 //! pub struct bool;
 //!
 //! impl QueryId for bool {
-//!     const SOME_PROPERTY: core::primitive::bool = true;
+//!     const SOME_PROPERTY: ::core::primitive::bool = true;
 //! }
 //!
-//! # trait QueryId { const SOME_PROPERTY: core::primitive::bool; }
+//! # trait QueryId { const SOME_PROPERTY: ::core::primitive::bool; }
 //! ```
+//!
+//! We also used `::core` instead of `core`, because `core` can be
+//! shadowed, too. Paths, starting with `::`, are searched in
+//! the [extern prelude] since Edition 2018.
+//!
+//! [extern prelude]: https://doc.rust-lang.org/nightly/reference/names/preludes.html#extern-prelude
 
 #[stable(feature = "core_primitive", since = "1.43.0")]
 pub use bool;
@@ -40,7 +46,7 @@ pub use f32;
 #[stable(feature = "core_primitive", since = "1.43.0")]
 pub use f64;
 #[stable(feature = "core_primitive", since = "1.43.0")]
-pub use i128;
+pub use i8;
 #[stable(feature = "core_primitive", since = "1.43.0")]
 pub use i16;
 #[stable(feature = "core_primitive", since = "1.43.0")]
@@ -48,13 +54,13 @@ pub use i32;
 #[stable(feature = "core_primitive", since = "1.43.0")]
 pub use i64;
 #[stable(feature = "core_primitive", since = "1.43.0")]
-pub use i8;
+pub use i128;
 #[stable(feature = "core_primitive", since = "1.43.0")]
 pub use isize;
 #[stable(feature = "core_primitive", since = "1.43.0")]
 pub use str;
 #[stable(feature = "core_primitive", since = "1.43.0")]
-pub use u128;
+pub use u8;
 #[stable(feature = "core_primitive", since = "1.43.0")]
 pub use u16;
 #[stable(feature = "core_primitive", since = "1.43.0")]
@@ -62,6 +68,6 @@ pub use u32;
 #[stable(feature = "core_primitive", since = "1.43.0")]
 pub use u64;
 #[stable(feature = "core_primitive", since = "1.43.0")]
-pub use u8;
+pub use u128;
 #[stable(feature = "core_primitive", since = "1.43.0")]
 pub use usize;
